+++ src/Spice86.Core/Emulator/OperatingSystem/DosPathResolver.cs
@@ -15,18 +15,18 @@ internal class DosPathResolver {
     internal const char DirectorySeparatorChar = '\\';
     internal const char AltDirectorySeparatorChar = '/';
     private const int MaxPathLength = 255;
+    private const int DosMfnlength = 8;
+    private const int DosExtlength = 3;
+    private const int LfnNamelength = 255; // for the sanity check only
 
     private readonly DosDriveManager _dosDriveManager;
-    private readonly LocalFileSearchManager _localFileSearchManager;
 
     /// <summary>
     /// Initializes a new instance.
     /// </summary>
     /// <param name="dosDriveManager">The shared class to get all mounted DOS drives.</param>
-    /// <param name="localFileSearchManager">Provides cached file searches</param>
-    public DosPathResolver(DosDriveManager dosDriveManager, LocalFileSearchManager localFileSearchManager) {
+    public DosPathResolver(DosDriveManager dosDriveManager) {
         _dosDriveManager = dosDriveManager;
-        _localFileSearchManager = localFileSearchManager;
     }
 
     /// <summary>
@@ -218,8 +218,7 @@ internal class DosPathResolver {
             ReturnSpecialDirectories = false
         };
 
-        string[] matches = _localFileSearchManager.FindFilesUsingWildCmp(resolvedHostDir, lastSegment, options);
-        string? firstMatch = matches.FirstOrDefault();
+        string? firstMatch = FindFilesUsingWildCmp(resolvedHostDir, lastSegment, options).FirstOrDefault();
         return string.IsNullOrWhiteSpace(firstMatch) ? null : ConvertUtils.ToSlashPath(firstMatch);
     }
 
@@ -326,5 +325,128 @@ internal class DosPathResolver {
         return Directory
             .GetDirectories(hostPath, searchPattern)
             .Concat(Directory.GetFiles(hostPath, searchPattern));
+    }
+
+
+    public IEnumerable<string> FindFilesUsingWildCmp(string searchFolder, string searchPattern,
+        EnumerationOptions enumerationOptions) {
+        return Directory.EnumerateFileSystemEntries(searchFolder, "*", enumerationOptions)
+            .Where(path => WildFileCmp(Path.GetFileName(path), searchPattern));
+    }
+
+    public static bool WildFileCmp(string? filename, string? pattern) {
+        if (filename is null || pattern is null) {
+            return false;
+        }
+
+        return WildFileCmp(filename.AsSpan(), pattern.AsSpan());
+    }
+
+    private static bool WildFileCmp(ReadOnlySpan<char> sourceFilename, ReadOnlySpan<char> pattern) {
+        if (sourceFilename.Length > 0 && pattern.Length == 0) {
+            return false;
+        }
+
+        if (pattern.Length > LfnNamelength) {
+            return false;
+        }
+
+        // Fast path: exact case-insensitive match (covers common no-wildcard cases)
+        if (sourceFilename.Equals(pattern, StringComparison.OrdinalIgnoreCase)) {
+            return true;
+        }
+
+        // Skip “hidden” dot-files if the pattern uses wildcards (except "." / "..")
+        if (WildcardMatchesHiddenFile(sourceFilename, pattern)) {
+            return false;
+        }
+
+        // Uppercase once into fixed-size 8.3 stacks with space padding
+        Span<char> fileName = stackalloc char[DosMfnlength];
+        Span<char> fileExt = stackalloc char[DosExtlength];
+        Span<char> wildName = stackalloc char[DosMfnlength];
+        // wild ext needs an extra slot to check the 4th char like DOSBox
+        Span<char> wildExt = stackalloc char[DosExtlength + 1];
+
+        SplitTo83(sourceFilename, fileName, fileExt, out _);
+        SplitTo83(pattern, wildName, wildExt, out int wildExtLength);
+
+        // ---- NAME compare (no early '*' accept; '*' just ends name matching) ----
+        if (CompareSegment(fileName, wildName, DosMfnlength) == false) {
+            return false;
+        }
+        // the original code tolerated extra chars in pattern name beyond 8 unless it's a non-'*' check byte; our fixed-size truncation mirrors that behavior
+
+        // ---- EXT compare (early '*' accept) ----
+        return CompareSegment(fileExt, wildExt, DosExtlength) switch {
+            true => true,
+            false => false,
+            // If wild ext has a 4th char, and it's not '*', reject (DOSBox-like behavior)
+            _ => wildExtLength <= DosExtlength || wildExt[DosExtlength] == '*'
+        };
+    }
+
+    /// <summary>
+    /// Common 8.3 segment compare.
+    /// </summary>
+    /// <param name="filenameOrExt">The segment to compare against.</param>
+    /// <param name="pattern">The pattern to match.</param>
+    /// <param name="length">Maximum number of characters to compare.</param>
+    /// <returns>
+    /// true - definitively accept (only possible with earlyStarAcceptsTrue and '*' seen)
+    /// false - mismatch 
+    /// null - matched the segment fully (or name-stopped at '*'), no final decision
+    /// </returns>
+    /// <remarks>
+    /// '?' matches any char (including space padding).
+    /// If '*' is encountered:
+    /// - For extension compare (earlyStarAcceptsTrue=true): returns true immediately
+    /// - For name compare (earlyStarAcceptsTrue=false): treats as "stop comparing here" and returns null
+    /// </remarks>
+    private static bool? CompareSegment(ReadOnlySpan<char> filenameOrExt, ReadOnlySpan<char> pattern, int length) {
+        for (int i = 0; i < length; i++) {
+            char patternChar = pattern[i];
+            if (patternChar == '*') {
+                return true;
+            }
+
+            if (patternChar != '?' && patternChar != filenameOrExt[i]) {
+                return false;
+            }
+        }
+
+        return null;
+    }
+
+    private static void SplitTo83(ReadOnlySpan<char> file, Span<char> targetFileName, Span<char> targetFileExt,
+        out int extLength) {
+        targetFileName.Fill(' ');
+        targetFileExt.Fill(' ');
+
+        int dotPos = file.LastIndexOf('.');
+        ReadOnlySpan<char> fileNameRaw = dotPos >= 0 ? file[..dotPos] : file;
+        ReadOnlySpan<char> fileExtRaw =
+            dotPos >= 0 && dotPos + 1 < file.Length ? file[(dotPos + 1)..] : ReadOnlySpan<char>.Empty;
+        ToUpperCopy(fileNameRaw[..Math.Min(fileNameRaw.Length, targetFileName.Length)], targetFileName);
+        ToUpperCopy(fileExtRaw[..Math.Min(fileExtRaw.Length, targetFileExt.Length)], targetFileExt);
+        extLength = fileExtRaw.Length; // actual (untruncated) length for the DOSBox-style 4th-char check
+    }
+
+    private static bool WildcardMatchesHiddenFile(ReadOnlySpan<char> fileName, ReadOnlySpan<char> wildcard) {
+        if (fileName.IsEmpty) {
+            return false;
+        }
+
+        if (wildcard.IndexOfAny('?', '*') < 0) {
+            return false;
+        }
+
+        return fileName.Length >= 5 && fileName[0] == '.' &&
+               !fileName.Equals(".", StringComparison.Ordinal) &&
+               !fileName.Equals("..", StringComparison.Ordinal);
+    }
+
+    private static void ToUpperCopy(ReadOnlySpan<char> src, Span<char> dst) {
+        src.ToUpperInvariant(dst);
     }
 }
